Add files via upload
Finished part 4 of the tutorurial: http://www.deepideas.net/deep-learning-from-scratch-v-multi-layer-perceptrons/

binary_data_generator.py: generates data for the other programs. It is useful, but it was not made to be versatile, it only attends the needs of this especific example.

computational_graphs.py: library with all the functions that actually do the magic. To be honest, I still don't really understand how it works!

multi_class_perceptron.py: labels multiple data but does not yet train the perceptron.

perceptron_training.py: acctually trains the perceptron. 

training_criterion.py: an introduction prior to perceptrion_training.py. it just labels one piece of data on its likelyhood of been blue or red.